Move stringIsISO8601() from Kernel to Types::Core and rewrite.
Makes the function accessible to Types::Core::DateAndTime. The function
was originally using Poco which is not linked to Types. We now use
boost::regex instead.

Re #20969
