+++ Framework/DataHandling/src/LoadLog.cpp
@@ -6,15 +6,14 @@
 #include "MantidAPI/FileProperty.h"
 #include "MantidDataObjects/Workspace2D.h"
 #include "MantidKernel/ArrayProperty.h"
-#include "MantidKernel/DateAndTimeHelpers.h"
 #include "MantidKernel/Glob.h"
 #include "MantidKernel/LogParser.h"
 #include "MantidKernel/PropertyWithValue.h"
 #include "MantidKernel/Strings.h"
 #include "MantidKernel/TimeSeriesProperty.h"
 #include "MantidKernel/make_unique.h"
+#include "MantidTypes/Core/DateAndTimeHelpers.h"
 
-#include <MantidKernel/DateAndTimeHelpers.h>
 #include <Poco/DateTimeFormat.h>
 #include <Poco/DateTimeParser.h>
 #include <Poco/DirectoryIterator.h>
@@ -469,13 +468,12 @@ bool LoadLog::isAscii(const std::string &filename) {
 }
 
 /**
- * Check if first 19 characters of a string is date-time string according to
- * yyyy-mm-ddThh:mm:ss
+ * Check if the string conforms to the ISO8601 standard.
  * @param str :: The string to test
  * @returns true if the strings format matched the expected date format
  */
 bool LoadLog::isDateTimeString(const std::string &str) const {
-  return DateAndTimeHelpers::stringIsISO8601(str.substr(0, 19));
+  return Types::Core::DateAndTimeHelpers::stringIsISO8601(str.substr(0, 19));
 }
 
 /**
