+++ Framework/Kernel/inc/MantidKernel/DateAndTimeHelpers.h
@@ -11,8 +11,6 @@ namespace DateAndTimeHelpers {
 MANTID_KERNEL_DLL Types::Core::DateAndTime
 createFromSanitizedISO8601(const std::string &date);
 
-MANTID_KERNEL_DLL bool stringIsISO8601(const std::string &date);
-
 MANTID_KERNEL_DLL std::string
 verifyAndSanitizeISO8601(const std::string &date, bool displayWarnings = true);
