+++ Framework/Types/CMakeLists.txt
@@ -1,19 +1,20 @@
-# This is not a module, there are no source files. Types in this folder must be header-only!
-
 set ( SRC_FILES 
   src/Core/DateAndTime.cpp
+  src/Core/DateAndTimeHelpers.cpp
   src/Event/TofEvent.cpp
 )
 
 set ( INC_FILES
   inc/MantidTypes/Core/DateAndTime.h
-  inc/MantidTypes/SpectrumDefinition.h
+  inc/MantidTypes/Core/DateAndTimeHelpers.h
   inc/MantidTypes/Event/TofEvent.h
+  inc/MantidTypes/SpectrumDefinition.h
 )
 
 set ( TEST_FILES
-  SpectrumDefinitionTest.h
   DateAndTimeTest.h
+  DateAndTimeHelpersTest.h
+  SpectrumDefinitionTest.h
   TofEventTest.h
 )
